Adding reactive behavior for porject section
This commit provides the js code for the project slideshow and text
update. Pressing the left button will append the first element in the
projects htmlcollections at the end (results in an update). Clicking the
right button will append the first two elements at the end (results in
update/ right slide)

Changes to be committed:
	modified:   index.html
	modified:   main.js
	modified:   style.css